Progress un ENCUESTAME-191

+++ encuestame-business/src/main/java/org/encuestame/business/service/imp/IFrontEndService.java
@@ -105,5 +105,5 @@ public interface IFrontEndService extends ServiceOperations {
      * @param limit
      * @return
      */
-    Integer getHashTagFrecuency(final Long hashTagId, final Integer limit);
+    Long getHashTagFrecuency(final Long hashTagId, final Integer limit);
 }
